# Simplify up interleave.