--- a/cytoolz/itertoolz.pxd
+++ b/cytoolz/itertoolz.pxd
@@ -36,9 +36,8 @@ cdef object c_merge_sorted(object seqs, object key=*)
 
 cdef class interleave:
     cdef list iters
-    cdef list newiters
     cdef Py_ssize_t i
-    cdef Py_ssize_t n
+    cdef int active
 
 
 cdef class _unique_key: