Simplified the implementation of interleave and made it slightly faster.

```python
seqs = [range(1000) for i in range(100)]
%timeit list(interleave(seqs))
```
old: 5.07 ms ± 240 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
new: 4.03 ms ± 45.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)